requirements.txt
llama-index==0.10.32
openai==1.14.2
python-dotenv==1.0.1
tqdm==4.66.2
PyMuPDF==1.23.22
pypdf==4.0.1
fastapi==0.110.0
uvicorn==0.29.0
